Implementando metodo para buscar usuario

+++ mscadastro/src/main/java/com/devpro/mscadastro/service/UsuarioService.java
@@ -7,6 +7,8 @@ import com.devpro.mscadastro.repository.UsuarioRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @AllArgsConstructor
 @Service
 public class UsuarioService {
@@ -14,12 +16,21 @@ public class UsuarioService {
     private UsuarioRepository usuarioRepository;
     private UsuarioMapper usuarioMapper;
 
-    public void cadastrarUsuario(UsuarioDto usuarioDto) {
+    public String cadastrarUsuario(UsuarioDto usuarioDto) {
         Usuario usuario = usuarioMapper.converteToUsuario(usuarioDto);
         usuarioRepository.save(usuario);
+        return "Usuario cadastrado com sucesso !";
     }
 
-    public void deletarUsuario(Long id) {
-        usuarioRepository.deleteById(id);
+    public String deletarUsuario(Long id) {
+        if (buscarUsuarioPorId(id).isPresent()) {
+            usuarioRepository.deleteById(id);
+            return "Usuario deletado com sucesso !";
+        }
+        return "Usuario não encontrado";
+    }
+
+    private Optional<Usuario> buscarUsuarioPorId(Long id) {
+        return usuarioRepository.findById(id);
     }
 }
